+++ src/cmds/test-dependant.js
@@ -39,10 +39,15 @@ module.exports = {
         return deps
       },
       default: {}
+    },
+    scriptName: {
+      describe: 'The script name to run',
+      type: 'string',
+      default: 'test'
     }
   },
   /**
-   * @param {{ repo: string; branch: string; deps: any; }} argv
+   * @param {{ repo: string; branch: string; deps: any; scriptName: string}} argv
    */
   async handler (argv) {
     const cmd = require('../test-dependant')