+++ test/dependants.js
@@ -87,7 +87,7 @@ describe('dependants', function () {
       expect(output.stdout).to.include(`${diff}-dependency-version=1.0.1`)
     })
 
-    it('should fail to test a project that does not depend on the deps we are overridding', async () => {
+    it('should fail to test a project that does not depend on the deps we are overriding', async () => {
       const diff = `derp-${Math.random()}`
 
       await expect(execa(bin, ['test-dependant', dirs.project, '--deps=it-derp@1.0.1'], {
@@ -95,6 +95,22 @@ describe('dependants', function () {
           DISAMBIGUATOR: diff
         }
       })).to.eventually.be.rejectedWith(/Module project does not depend on it-derp/)
+    })
+
+    it('should run a given script in a regular project', async () => {
+      const diff = `derp-${Math.random()}`
+
+      const output = await execa(
+        bin,
+        ['test-dependant', dirs.project, '--deps=it-all@1.0.1 --script-name=another-test'],
+        {
+          env: {
+            DISAMBIGUATOR: diff
+          }
+        }
+      )
+      expect(output.stdout).to.include(`${diff}-dependency-version=1.0.0`)
+      expect(output.stdout).to.include(`${diff}-dependency-version=1.0.1`)
     })
   })
 
@@ -106,6 +122,18 @@ describe('dependants', function () {
     it('should test a monorepo', async () => {
       const diff = `derp-${Math.random()}`
       const output = await execa(bin, ['test-dependant', dirs.monorepo, '--deps=it-all@1.0.1'], {
+        env: {
+          DISAMBIGUATOR: diff
+        }
+      })
+
+      expect(output.stdout).to.include(`${diff}-dependency-version=1.0.0`)
+      expect(output.stdout).to.include(`${diff}-dependency-version=1.0.1`)
+    })
+
+    it('should run a given script in a monorepo project', async () => {
+      const diff = `derp-${Math.random()}`
+      const output = await execa(bin, ['test-dependant', dirs.monorepo, '--deps=it-all@1.0.1 --script-name=another-test'], {
         env: {
           DISAMBIGUATOR: diff
         }